TESTrun: Implement --passed.
Gisle Vanem in GH #1466 comments that sometimes it is useful to see an
explicit list of all passed tests.  In this case perhaps it would also
be useful to see how long each of those took to complete, in order to
decide what timeout value to set.

To implement that, introduce a new time_test_command() subroutine, which
returns the exit status code and possibly the duration if the user
requested it through the new command-line option.  In run_reject_test()
and validate_stdout_test() use the new subroutine and for passed tests
feed the duration, if available, to result_passed().  Extend the latter
to return the duration to test_and_report(); there print a table of
passed tests details if available, but do not expect any details to be
present in every case.  Add T min/avg/max to the final statistics.