TESTrun: Implement --config. [skip appveyor]
Gisle Vanem in GH #1466 further comments that config.h may be using DOS
newlines, also that the script is not a C preprocessor and its
interpretation of config.h sometimes requires verification, so provide
an easy way to see what it managed to parse.

Return the hash from read_config_h() and in TESTrun print the contents
if the user specifies the command-line option.  This also makes it easy
to notice that the hash initialization in read_config_h() isn't correct,
which adds stray data into the result, make that right.  Fix the quoted
string matching regexp to match empty strings (i.e. PACKAGE_URL) and
merge the two regexps into a more readable multiline extended one that
disregards any stray CR/LF characters.  Read strings together with the
quotes because this way the output of --config looks more consistent
with config.h and the regexp is much simpler.