TESTrun: Accommodate Windows hosts file better.
Gisle Vanem confirms the file works on Windows, as far as pcap_compile()
sees it, so let's try testing this as well.

Since Windows in its hosts file ignores link-local IPv6 addresses that
do not have a Windows zone ID suffix, the easiest way to use the same
hosts file contents on all OSes seems to be renumbering the test IPv6
address to a non-link-local prefix, so try that.  In skip_no_hosts()
either refer to a missing environment variable or use a Windows-specific
path to the hosts file.  Since the latter does not fit into the output
format, in skip_unless_file_contains_lines() use the basename in the
skip message.