pcap-filter(7): Document "ip broadcast" better. [skip ci]
Explain where the netmask comes from and how it works in this primitive.
Make it clear which address bits it compares and which does not, provide
an example and an alternative syntax.  Remove "...looks up the subnet
mask on the interface on which the capture is being done", which belongs
to tcpdump(1).