Filter SocketCAN frames in userland if necessary.
On Linux little-endian hosts the first four bytes of a SocketCAN frame
are the opposite byte order at the kernel side of the socket, so a
filter program that accesses any of these bytes needs either to account
for the byte order difference (if possible) or to run in userland.
Implement this logic in fix_dlt_can_socketcan() and call it from
fix_program() if necessary.  See also GitHub bug report #1247.