WIP/RFC: add support for manual link layer offset changes
A few pcap-filter expressions implicitly change the link layer offset.
There are other less common headers without their own keywords that
might require skipping over some bytes in a link layer adjacent header
to get to higher layer network headers. This commit adds an "offset"
term that takes a + or - value to adjust the link layer offets in the
same way VLAN does. This allows for something like `ethertype 0x8926
and offset +6 and vlan 14 and ip host 192.0.2.14` to skip over a vntag
header, process a VLAN header, and look for an IPv4 host.