pcap-filter(7): Refine qualifiers and primitives.
In the introduction mention that an ID is not limited to a name or a
number, add "proto" to the list of type qualifiers and give an example,
mention that a proto qualifier and the "proto" type qualifier are not
the same, add more entries to the list of valid protocols, give more
examples, say that "proto" implies "ip or ip6" and "net" also implies
"ip6".

Ibid., add a paragraph to explain how the qualifiers combine (or not) in
the regular primitive syntax; in IEEE 802.11 header particulars remove
the comment about RA and TA because the current implementation does in
fact support these, remove the comment about BSSID because the list of
"wlan" primitives represents what is supported, everything else is not.
Ibid., add ARP/RARP header particulars, rearrange a couple paragraphs
to discuss one thing at a time and provide a few more examples of the
irregular primitive syntax.

For each of the primitives below merge "dst", "src" and "src or dst"
forms into one section using a uniform style and in each section give a
complete list of valid qualifiers and related details:
* "host HOSTNAMEADDR"
* "ether host ETHERNAMEADDR"
* "net NETNAMEADDR"
* "net NETADDR mask NETMASK"
* "net NETADDR/LEN"
* "port PORTNAMENUM"
* "portrange PORTNAMENUM1-PORTNAMENUM2"
* "decnet HOST"
* "link HOST" with regard to ARCnet

Update "gateway HOST" along the same lines to make it clear it matches
ARP and RARP too by default.  For "ether host ETHERNAMEADDR", "ether
broadcast" and "ether multicast" mention that Ethernet is not the only
supported network type.  For "net NETNAMEADDR" spell the CIDR notation
of every netmask.  Ibid., add a reference to "net NETADDR/LEN".  Idem
for "net NETADDR mask NETMASK", also discuss the argument syntax and
clarify that the netmask can be any 32-bit value.  For "portrange" make
it clear that the two values can be in any order and may be equal.  Add
a note about ARP and RARP to the "bugs" section.