man: Document devices, interfaces and "any" better.
In pcap_create(3PCAP) and pcap_open_live(3PCAP) spell device/interface
and replace an explanation of the difference with a reference to
pcap_findalldevs(3PCAP).

In the latter man page remove the text that used to say "does not have
sufficient privileges... those devices will not appear" since libpcap
1.1.1 because it does not seem to match even what old libpcap versions
implemented (e.g. skip only if IFF_UP is not set in fad-gifc.c).  Add a
paragraph to explain the difference between listing devices and
activating a device in the current implementation.  Explain that a
capture device is a generic case of a network interface and how the
"any" pseudo-interface relates with network interfaces.  Show that a
device name can be used with pcap_create() too.

This resolves GitHub bug report #1055.