TESTrun: Convert test definitions from HoH to AoH.
The following is valid Perl code (although Python implements the same
semantics for its dictionaries too):

my %h = (key1 => 1, key1 => 2);

The problem here is, if the developer makes a mistake and uses the same
key to define more than one test, neither Perl interpreter nor the
duplicate label check will report an error, and at least one test will
be a silent no-op.  Make the test name a key of the second-level hash,
convert the top-level hashes into arrays, add checks for "name" in the
hash and do not sort tests within the three groups anymore -- just use
the order of the arrays.