Replace gen_check_802_11_data_frame() with gen_mcmp().
In the FC0 byte of the IEEE 802.11 header instead of testing separately
one bit for being set and another bit for being unset test one value
using a 2-bit bitmask.  For each of the five DLTs that used the function
this change trades two instructions for two other instructions, also for
DLT_IEEE802_11 and DLT_PRISM_HEADER it eliminates duplicate loads.
Another effect is that for the five DLTs "llc" and "wlan type data" now
generate exactly the same filter program.  Update the accept tests to
reflect that; the new filter programs have the same effect as before, so
the apply tests do not change.