Enable "[wlan] dir/type/subtype" for DLT_PPI.
For DLT_PPI "llc", which is equivalent to "wlan type data", produces a
correct filter program, so there is no reason not to support "wlan type"
and "wlan subtype" which test the same packet data using a value and a
mask.  Likewise, "wlan dir" produces a correct filter program, so enable
it too.  Add/update tests as required to cover the new code paths.

The file http_PPI.pcap is a reduced version of http_PPI.cap from the
Wireshark sample captures collection.