pcap-filter(7): Document protocol names better. [skip ci]
Add a table with a summary of all valid protocol names and related
usage.  Add a description for "isis proto NUM".  List "isis", "ppp" and
"slip" as valid proto qualifier cases, list "protochain" and "gateway"
as valid type qualifier cases, list "link", "ppp" and "slip" as valid
link-level aliases, remove "gateway" and "wlan" from the examples of
irregular primitives.  In the lists of qualifiers try to put the most
frequently used first.  When mentioning irregular primitives, remember
to say "relation" because arithmetic expressions evaluate to an integer,
not a Boolean.