Use the correct IS-IS PDU type offset for DLT_C_HDLC.
The very first octet of an IS-IS PDU header is "Intradomain Routeing
Protocol Discriminator".  The default case in gen_proto() reads it from
offset 0 of the link-layer payload for most DLTs and from offset 1 for
DLT_C_HDLC, which works as expected.  However, the Q_ISIS case always
uses the same offset, so IS-IS PDU type keywords fail to match packets
correctly for DLT_C_HDLC.  Copy the workaround from Q_ISO to address
that and update the tests.