Use the correct bit mask for IS-IS PDU type value.
The IS-IS PDU Type field is an unsigned integer in the least significant
5 bits, and the most significant three bits are reserved, therefore when
"isis proto" tests the field value, it must use a bit mask to ignore the
reserved bits.  This notably inflates the filter program because the
optimizer currently detects no-op duplicate loads, but does not detect
no-op duplicate (ld, and) pairs.  The apply tests do not change because
they have packets with the reserved bits set to zero.