set break_loop in pcap.c when PCAP_ERROR_BREAK is returned
The code is the same everywhere anyway, and there are just a few callers
of read_op that need to deal with this. So let the caller handle it.

This should be a bit nicer overall and will simplify the implementation
to permit breaking the loop in a way that also flushes out the queue.