add pcap_breakloop_flush API and implement it on linux
It is useful to get all the packets up to the current point in time. Due
to input buffering this may not be the case. On Linux with tcpdump, one
can get up to a second worth of packets that are not processed.

Add a new pcap_breakloop_flush and set the internal break_loop variable
to separate values depending on which version of pcap_breakloop was
used. Then, on Linux, keep processing packets without polling if the
break_loop variable is set to PCAPINT_BREAK_FLUSH, but abort immediately
in the PCAPINT_BREAK_IMMEDIATE case.