Mind all-zeroes IPv4 netmask in gen_host().
When the IPv4 netmask is /0, gen_mcmp() returns an equivalent of either
true or false.  In the context of gen_host() it is clear when it is
going to be true, so take a shortcut and return a block with fewer
statements.

Now for "ip dst net 0.0.0.0/0" the unoptimized and the optimized filter
programs are the same (without the always-true comparison).