Generate cleaner code for unqualified "port" and "portrange".
For a given AF (ip/ip6) and direction (src/dst), the only difference
between a TCP port check, a UDP port check and an SCTP port check is
the protocol number, all other logistics is the same.  Lose some
duplicate C code, in the remaining code avoid generating some of the
duplicate instructions (see also commit 2559282).  This way the
unoptimized filter programs are notably shorter and the optimized filter
programs are exactly the same.