TESTrun: Add more reject tests with pcap_ether_hostton().
pcap_ether_hostton() returns an allocated 48-bit binary MAC address,
which ought to be freed no matter if the attempt to use it was
successful or not.  The latter case stands for several code paths that
start at gen_scode(), invoke pcap_ether_hostton() and eventually
bpf_error(); each of these ought to arrange a means to free the memory.
Some of these code paths are already covered with tests, add tests to
cover the remaining code paths and make the problem space completely
visible to Valgrind.