Use inet_pton() in gen_mcode6().
The function needs to translate a [well-formed] IPv6 address string to a
struct in6_addr.  Although pcap_nametoaddrinfo() in fact does that, it
overshoots the problem space unnecessarily and requires managing of the
associated freeaddrinfo() call.  Also the "resolved to multiple address"
condition is never true because in this case the string is an address,
not a hostname, and cannot translate to more than one binary address.

Switch to inet_pton(), which fits the problem space much better, and
verify that it has not failed (very unlikely, but still).