WIP: Simplify gen_gateway().
In gen_scode() case Q_GATEWAY delegate the remaining memory management
and error handling away; now if gen_gateway() returns, it always returns
a non-NULL value.  In gen_gateway() handle all errors as early as
possible; instead of testing whether ai_addr is NULL test whether
ai_family is set to the AF of interest, when that's the case, ai_addr is
not NULL (this is how gen_scode() implements case Q_HOST).