dlpi: improve some error messages.
Try to provide a more precise error indication when we have a problem
opening a DLPI device.

This should, among other things, handle the "It might also be worth
having pcap_activate() return PCAP_ERROR_CAPTURE_NOTSUP for that case,
and if this is running in a zone, provide a message suggesting that the
zone might not allow that." comment in
https://github.com/the-tcpdump-group/libpcap/issues/1425#issuecomment-2569956919.

Expand some comments while we're at it.

(cherry picked from commit c0c6e3030a7d947d572246365e714b2ef02a2953)