dpdk: suppress -Wformat-truncation warnings.
This doesn't fix them - nothing short of changing the way we generate
error message strings will do that - but at least it means we don't get
annoyed by them, and it also marks places that would need to be changed
if we change the way we generate those error message strings. See #1544
for an issue for making those changes.

This addresses part of #1543.