TESTrun: Get SKF_AD_VLAN_TAG_PRESENT right. [skip appveyor]
As discussed in GH#1554, HAVE_DECL_SKF_AD_VLAN_TAG_PRESENT appears in
config.h only when pcap-linux.c needs to be compiled, which is a subset
of all cases when the OS is Linux.  Specifically, "--with-pcap=null" is
one of the cases that do not define HAVE_DECL_SKF_AD_VLAN_TAG_PRESENT.
At the same time, skip_config_have_decl() expects the given macro to be
defined and breaks.

Redo the affected skip conditions to use the recently renamed
skip_config_not_def1() instead and to disregard the OS altogether.
Delete the now unused skip_config_have_decl().