Fix "(arp|rarp) host NAME" to ignore IPv6 quietly.
For these two forms of the predicate, when a hostname resolves in both
IPv4 and IPv6, the IPv6 part is supposed to be irrelevant, but it causes
an unexpected failure (GH bug #1512).  Also, when a hostname does not
resolve in IPv4 and resolves in IPv6, the error message is supposed to
be correct, but it is not:

$ filtertest EN10MB 'arp host eth-ipv4-ipv6.host123.libpcap.test'
filtertest: 'arp' is not a valid qualifier for 'ip6 host'
$ filtertest EN10MB 'rarp host eth-ipv4-ipv6.host123.libpcap.test'
filtertest: 'rarp' is not a valid qualifier for 'ip6 host'
$ filtertest EN10MB 'arp host eth-noipv4-ipv6.host123.libpcap.test'
filtertest: 'arp' is not a valid qualifier for 'ip6 host'
$ filtertest EN10MB 'rarp host eth-noipv4-ipv6.host123.libpcap.test'
filtertest: 'rarp' is not a valid qualifier for 'ip6 host'

In gen_scode() case Q_HOST get the IPv6 address skip condition right to
fix this:

$ filtertest EN10MB 'arp host eth-ipv4-ipv6.host123.libpcap.test'
(000) ldh      [12]
(001) jeq      #0x806           jt 2	jf 7
(002) ld       [28]
(003) jeq      #0xa141e28       jt 6	jf 4
(004) ld       [38]
(005) jeq      #0xa141e28       jt 6	jf 7
(006) ret      #262144
(007) ret      #0
$ filtertest EN10MB 'rarp host eth-ipv4-ipv6.host123.libpcap.test'
(000) ldh      [12]
(001) jeq      #0x8035          jt 2	jf 7
(002) ld       [28]
(003) jeq      #0xa141e28       jt 6	jf 4
(004) ld       [38]
(005) jeq      #0xa141e28       jt 6	jf 7
(006) ret      #262144
(007) ret      #0
$ filtertest EN10MB 'arp host eth-noipv4-ipv6.host123.libpcap.test'
filtertest: unknown host 'eth-noipv4-ipv6.host123.libpcap.test' for
  specified address family
$ filtertest EN10MB 'rarp host eth-noipv4-ipv6.host123.libpcap.test'
filtertest: unknown host 'eth-noipv4-ipv6.host123.libpcap.test' for
  specified address family

In pcap-filter(7) explain the address family semantics better in the
description of "host NAME".  Update the tests.