Fix error messages about 32-bit integer overflow.
Since commit a5cac25 stoulen() takes a pointer to a string and a
length of the string and uses the length to report a problem about the
string.  However, by that time it has already modified the length, so
the error message does not match the input:

$ filtertest RAW '123456789012345'
filtertest: number 1234 overflows 32 bits
$ filtertest RAW '12345678901234567890'
filtertest: number 123456789 overflows 32 bits
$ filtertest RAW '123456789012345678901234567890'
filtertest: number 1234567890123456789 overflows 32 bits
$ filtertest RAW '0123456701234'
filtertest: number  overflows 32 bits
$ filtertest RAW '0x100000000'
filtertest: number  overflows 32 bits

Consequently, if more than one number in the filter expression is out of
range, it can be not immediately obvious which one it is:

$ filtertest RAW '4294967296 != 4294967297'
filtertest: number  overflows 32 bits

To fix this, in stoulen() keep a copy of the original string length and
use it for the error reporting.  Ibid., factor the format string out and
show the base in the message.  Simplify the forward declaration.  Add
two comments to outline the clash space between certain MAC-48 addresses
and certain octal integers.  Document this syntax peculiarity in the man
page and add a few basic reject tests.