TESTrun: Refine comments with gen_scode() code paths.
Unlike gen_ncode(), gen_mcode() and gen_mcode6(), there is exactly one
code path into gen_scode(), namely, through an ID in the grammar.  This
way, when a comment says "ID -> gen_scode()", it states the obvious, and
when it says "pnum -> gen_scode()", it is wrong, so eliminate this noise
to improve the signal.  Add the missing gen_host46_byname() to a couple
code paths.