TESTrun: Eliminate some gaps in "host" reject tests.
The existing tests could not detect the recently fixed DECnet bugs
because the definitions in %pqual_features do not distinguish between
different forms of the ID.  Then the subsequent foreach loop that
generates invalid "host" primitives does not generate any tests for
protocols that support a form of "host" no matter which forms of ID are
valid and which are not.  Then the resulting reject tests have a number
of gaps, including IPv4 IDs in DECnet primitives.

To fix that, in %pqual_features replace the generic "host" with seven
more specific host/net features; in the loop skip the valid permutations
only and include MAC-48, ARCnet and DECnet IDs.  The new test space
supersedes a number of earlier reject tests, so remove these.  Document
the current code paths in the comments and introduce more helper
functions.