rpcapd-config(5): Document ActiveClient and PassiveClient better.
For ActiveClient reword the description for clarity.  For PassiveClient
remove all mentions of port numbers to match what fileconf_read()
implements, which in turn matches the behaviour implemented and
documented for the -l command-line flag, which in turn is the most
sensible behaviour because the source TCP port is almost certainly
ephemeral.