pcap-filter(7): IPv4-mapped IPv6 is valid syntax.
Moreover, it is valid in all contexts of an IPv6 address, so move the
definition to the place of first use ("host") and refer to inet_pton(3),
which does the parsing.  Later on just refer to the earlier definition.
Add IPv4-mapped aliases to the existing IPv6 accept tests, and a few
reject tests.