In gen_scode() always leave the proto qualifier intact.
The only effect of the conditional override that expands Q_DEFAULT to
Q_IP and Q_IPV6 (but not Q_ARP or Q_RARP) used to be absence of the ARP
and RARP dead branches in the unoptimized filter program for a number of
DLTs.  Since this is now a particular case of the generic AND/OR
reduction, remove the custom code -- this produces exactly the same
filter programs, so there is no change in any tests.