Get BPF_K and BPF_X right in gencode.c.
Apparently, earlier I made an error in commit 87b9146, in that it is not
right to replace five instances of a BPF_JMP|BPF_JEQ|BPF_X block with
calls to gen_jmp(): gen_jmp() uses JMP() to produce a BPF_K branch
instruction, which takes a constant for the "#k", but a BPF_X branch
instruction does not take a constant.

This error did not manifest in practice because BPF_K == 0x00 and BPF_X
== 0x08, so passing BPF_X to gen_jmp() by coincidence worked as before.
To make it obvious which variety of a branch block the code means to
produce, replace gen_jmp() with gen_jmp_k() and gen_jmp_x(), each of
which does its own correct thing.

Whilst prototyping the above bug fix, in a few cases I had confused
new_stmt() with gen_jmp() and omitted a mandatory value from the ORed
bitmasks, which was not obvious until the tests failed.  To make such
mistakes compile errors, have JMP() take two arguments and use the
macro instead of composing the bitmasks in place.