Do not special-case BPF_JEQ in gen_relation_internal().
The function stores one operand in A and the other in X.  For BPF_JGT
and BPF_JGE it generates "jgt x" and "jge x" respectively, which works
exactly as required.  However, for BPF_JEQ it subtracts X from A, then
compares A with 0.  Although the immediate effect of this is the same as
that of "jeq x", the former requires an additional instruction.  Also it
does not preserve the original value of the operand in A, which may get
in the way of subsequent optimizations; indeed, optimizer.c:opt_peep()
folds "sub x; jeq #0" into "jeq x".

Thus in gen_relation_internal() use one gen_jmp_x() invocation for all
cases.  In 13 accept blocks this eliminates from the unoptimized program
one instruction per every instance of "EXPR1 == EXPR2" and
"EXPR1 != EXPR2".