TESTrun: Refine a few assorted accept blocks.
In MTP tests, for example, "sls (3)" tests a relevant code path, but
"sls (3 or 3)" tests effects of the optimizer on a code path that is
already covered with another test.  This causes unnecessary differences
in the unoptimized program, so remove such aliases and expect the short
program to be the unoptimized program.  Idem ATM tests.  For the length
tests keep all the aliases, but split these into separate tests on
similar grounds.  Use the 00 prefix to test octal for IPv4 and fix a
comment.