TESTrun: Test proto-unqualified "(host|net)" better.
The existing tests that mix "ip host IPV4ADDR" and "host IPV4ADDR" etc.
on DLT_RAW and declare the same optimized filter program do not cover
the problem space well enough.  One problem with that is such blocks
cannot declare an unoptimized filter program because in the unoptimized
space the proto-qualified form does not contain any ARP/RARP branches,
but proto-unqualified form contains them as unreachable instructions.
Another problem is that this way it is impossible to verify that the
proto-unqualified form matches ARP and ARP under right conditions.

Thus move the proto-unqualified aliases to their own test blocks and use
a DLT that supports all the implied protocols, and many others -- this
way the test proves that the implicit set of protocols extends as far as
it is supposed to extend, and no further.  Limit the new blocks to
lowercase hostnames only, and repeat a quarter (with a fixed dir
qualifier) of the new blocks for DLT_RAW to cover the elimination of ARP
and RARP in greater detail for the next commits.  This adds the
following accept test blocks:

* net_ipv4addr_16bit_IP_OVER_FC (from ip_net_addr)
* src_net_ipv4addr_24bit_IP_OVER_FC (from ip_src_net_addr)
* dst_net_ipv4addr_8bit_IP_OVER_FC (from ip_dst_net_addr_8)
* src_and_dst_net_ipv4addr_12bit_IP_OVER_FC (new)
* src_and_dst_net_ipv4addr_12bit_RAW (new)

* net_ipv6addr_IP_OVER_FC (from ip6_net)
* src_net_ipv6addr_IP_OVER_FC (from ip6_src_net)
* dst_net_ipv6addr_IP_OVER_FC (new)
* src_and_dst_net_ipv6addr_IP_OVER_FC (new)
* src_and_dst_net_ipv6addr_RAW (new)

* net_name_EN10MB (new)
* src_net_name_EN10MB (from ip_src_net_name and ip_src_net_NAME)
* dst_net_name_EN10MB (from ip_dst_net_name and ip_dst_net_NAME)
* src_and_dst_net_name_EN10MB (new)
* src_and_dst_net_name_RAW (new)

* host_ipv4addr_FDDI (from ip_host_addr)
* src_host_ipv4addr_FDDI (from ip_src_host_addr)
* dst_host_ipv4addr_FDDI (from ip_dst_host_addr)
* src_and_dst_host_ipv4addr_FDDI (new)
* src_and_dst_host_ipv4addr_RAW (new)

* host_ipv6addr_IEEE802_11 (from ip6_host_addr)
* src_host_ipv6addr_IEEE802_11 (from ip6_src_host_addr)
* dst_host_ipv6addr_IEEE802_11 (from ip6_dst_host_addr)
* src_and_dst_host_ipv6addr_IEEE802_11 (new)
* src_and_dst_host_ipv6addr_RAW (new)

* host_noeth_ipv4_noipv6_EN10MB (from ip_host_name/ip_host_NAME)
* host_noeth_noipv4_ipv6_EN10MB (from ip6_host_name/ip6_host_NAME)
* host_noeth_ipv4_ipv6_EN10MB (new)
* src_host_noeth_ipv4_noipv6_EN10MB (from ip_src_host_name and
  ip_src_host_NAME)
* src_host_noeth_noipv4_ipv6_EN10MB (from ip6_src_host_name and
  ip6_src_host_NAME)
* src_host_noeth_ipv4_ipv6_EN10MB (new)
* dst_host_noeth_ipv4_noipv6_EN10MB (from ip_dst_host_name and
  ip_dst_host_NAME)
* dst_host_noeth_noipv4_ipv6_EN10MB (from ip6_dst_host_name and
  ip6_dst_host_NAME)
* dst_host_noeth_ipv4_ipv6_EN10MB (new)
* src_and_dst_host_noeth_ipv4_noipv6_EN10MB (new)
* src_and_dst_host_noeth_noipv4_ipv6_EN10MB (new)
* src_and_dst_host_noeth_ipv4_ipv6_EN10MB (new)
* src_and_dst_host_noeth_ipv4_noipv6_RAW (new)
* src_and_dst_host_noeth_noipv4_ipv6_RAW (new)
* src_and_dst_host_noeth_ipv4_ipv6_RAW (new)