Use inet_pton() directly to validate HID6 in scanner.l.
To validate a string to be a well-formed IPv6 address, an invocation of
getaddrinfo(3) with hints.ai_family set to AF_INET6 and hints.ai_flags
set to AI_NUMERICHOST effectively is the same as an invocation of
inet_pton(3), but involves an indirection with setup and teardown, so do
not over-engineer it.  See also commit 5acbaa8.