In gen_host46_byname() always process IPv4 before IPv6.
From the multi-platform testing point of view, getaddrinfo() can be an
unreliable external effect, which gets in the way of verifying, for
example, that "src host HOSTNAME" compiles exactly to the expected
filter program when the hostname resolves in both IPv4 and IPv6.  See
also GH #1574.

To reduce this problem space, iterate over the results of getaddrinfo()
up to two times: the first time process IPv4 addresses only, if present,
and the second time process IPv6 addresses only, if present.