For proto-unqualified "port" test for SCTP after TCP and UDP.
Point 3 of GH #1022 is that the current implementation of "port" in
gen_port_common() and gen_port6_common() tests for SCTP first, then for
TCP, then for UDP, but for a random packet SCTP is the least likely
protocol in this set.  Therefore make the SCTP match the last one.

As to which of UDP and TCP is more likely than the other in a random
packet, [1] discusses that in 1998-2008 TCP was substantially more
likely with no tendency for change, so until proven otherwise let's
suppose this still holds.  Therefore keep the UDP match after the TCP
match.

1: "Media Streaming Observations: Trends in UDP to TCP Ratio",
   pp. 147-162, vol. 3, No 3 & 4, year 2010, International Journal on
   Advances in Systems and Measurements
   https://www.iariajournals.org/systems_and_measurements/sysmea_v3_n34_2010_paged.pdf