Separate DECnet and IPv4 code paths better.
The only reason for gen_ncode() to be aware of DECnet is that DECnet
address syntax grammar-wise is a subset of IPv4 address syntax, so it is
a common entry point for both.  There is no reason for gen_host() to be
aware of DECnet because otherwise the function uses gen_hostop(), which
operates in IPv4 space.

To make it clear that the two code paths diverge early, make the
DECnet-specific parts of gen_ncode() and gen_host() a new function,
gen_dnhost().  In TESTrun add two reject tests and update various
comments to show the new code paths.

In pcapint_atodn() make the output argument 16-bit to match the address
size and return 1 on success for consistency with other functions that
parse network addresses.  Update translatetest accordingly.