Use bpf_error() to handle all unexpected dir qualifiers.
If a direction qualifier has been validated by the parser and a switch
block on the result has a case for every valid value (as is the case
now), the default case with abort() is dead code, otherwise it is a bug
in libpcap (as was the case before commit 4d6e12c and potentially could
be again in future) that is specific to the filter string.

The latter could or could not be a user input, but in any case instead
of terminating the process it would be more proper to report an ordinary
failure to compile the expression and to let the application handle it.
Thus in all switch blocks that take a direction qualifier handle both
the expected invalid values and any unexpected invalid values using
bpf_error().