Make some internal errors a pcap_compile() error.
Introduce a new error format string, this time for an internal error
caused by an invalid argument value, and use it instead of the 11
remaining abort() calls in gencode.c.  For example, in gen_wlanhostop()
the latter is an internal error (and dead code) because the switch block
enumerates all defined direction qualifiers, the same principle applies
to the other functions.  There are no tests for these code paths, which
are unreachable in the current implementation.