Mention DLT names in error messages as well.
Given a DLT description, it may be not immediately obvious which name
and number it stands for, which may obscure the meaning of an error
string.  Introduce a helper function to give both the name and the
description if possible, or at very least to tell the number, and use it
in gencode.c instead of pcap_datalink_val_to_description_or_dlt().
While at it, in gen_vlan() and gen_mpls_internal() instead of producing
a custom error string use fail_kw_on_dlt(), and re-jig the error string
in require_basic_bpf_extensions() to be easier to match in the tests.
In TESTrun update reject tests to verify the DLT name in error messages
that now include it.